modification

+++ Backend_ProjetTicket_OCP/src/main/java/com/example/Backend_ProjetTicket_OCP/Controller/SuiviTicketController.java
@@ -22,7 +22,7 @@ public class SuiviTicketController {
 
 
     @PostMapping("/assign/{ticketId}/to/{equipeId}")
-    public SuiviTicketDTO assignTicket(@PathVariable Long ticketId, @PathVariable Long equipeId) {
+    public SuiviTicketDTO assignTicket(@PathVariable int ticketId, @PathVariable Long equipeId) {
         return suiviTicketService.assignTicketToEquipe(ticketId, equipeId);
     }
